feat: add setting to override auto damage roll for clients (#1426)
This much needed update adds a new setting so that the GM's can choose to override the auto damage roll setting for players.  The auto damage roll setting is, currently, the only setting that has a 'client' rather than a 'world' scope.  Thus, each player can configure as they prefer.  With the new setting, "Override automatic damage rolls to GM Setting" - checking the box (to true) forces the auto damage setting to the state that the GM has chosen and removes the setting option for all players.  Note that a reload is required for all clients when changing this setting. This new setting might be useful where you'd like consistency in how the players roll damage or when there are users unfamiliar with FVTT.

This update also adds:
* Some style fixes for Monks Active Title Triggers as noted by @Vertigo (through experimental/debug setting to correct CSS styles).
* A new macro to roll Traveller characteristics courtesy of @JoshB.