fix: change settings that are fractions from numbers to string and add setting for critical rolls (#1441)
This release fixes an issue with a couple of settings that were prevented from being less than one by Chrome.  It also adds a new setting to override a natural 2 or 12 roll to be an "exceptional" failure/success (thanks to @ForjaSalvaje).  This setting does not alter the effect of the roll, just the degree of success description.  Also added are some updated Spanish translations courtesy of @ForjaSalvaje, too.