fix: allow SkillRoll Enricher to make characteristic  checks (#1513)
This release allows the '@SkillRoll' enricher to make characteristic checks.  Entering 'None' for the skill will now default to a characteristic check, e.g. `@SkillRoll[None/DEX 8+]{Make a DEX check}`.  Thanks to the suggestion of @krebs.

Also, a 'bug' is fixed that doesn't allow Target DM labels to have numbers. Thanks to @ZEE for finding.